Fix docs & README
