+++ docs/code/evals.rst
@@ -14,31 +14,31 @@ BLEU
 
 :hidden:`sentence_bleu`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.evals.sentence_bleu
+.. autofunction:: texar.torch.evals.sentence_bleu
 
 :hidden:`corpus_bleu`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.evals.corpus_bleu
+.. autofunction:: texar.torch.evals.corpus_bleu
 
 :hidden:`sentence_bleu_moses`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.evals.sentence_bleu_moses
+.. autofunction:: texar.torch.evals.sentence_bleu_moses
 
 :hidden:`corpus_bleu_moses`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.evals.corpus_bleu_moses
+.. autofunction:: texar.torch.evals.corpus_bleu_moses
 
 :hidden:`corpus_bleu_transformer`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.evals.corpus_bleu_transformer
+.. autofunction:: texar.torch.evals.corpus_bleu_transformer
 
 :hidden:`bleu_transformer_tokenize`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.evals.bleu_transformer_tokenize
+.. autofunction:: texar.torch.evals.bleu_transformer_tokenize
 
 :hidden:`file_bleu`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.evals.file_bleu
+.. autofunction:: texar.torch.evals.file_bleu
 
 
 Accuracy
@@ -46,9 +46,9 @@ Accuracy
 
 :hidden:`accuracy`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.evals.accuracy
+.. autofunction:: texar.torch.evals.accuracy
 
 
 :hidden:`binary_clas_accurac`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.evals.binary_clas_accuracy
+.. autofunction:: texar.torch.evals.binary_clas_accuracy
