+++ docs/code/losses.rst
@@ -9,23 +9,23 @@ MLE Loss
 
 :hidden:`sequence_softmax_cross_entropy`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.sequence_softmax_cross_entropy
+.. autofunction:: texar.torch.losses.sequence_softmax_cross_entropy
 
 :hidden:`sequence_sparse_softmax_cross_entropy`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.sequence_sparse_softmax_cross_entropy
+.. autofunction:: texar.torch.losses.sequence_sparse_softmax_cross_entropy
 
 :hidden:`sequence_sigmoid_cross_entropy`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.sequence_sigmoid_cross_entropy
+.. autofunction:: texar.torch.losses.sequence_sigmoid_cross_entropy
 
 :hidden:`binary_sigmoid_cross_entropy`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.binary_sigmoid_cross_entropy
+.. autofunction:: texar.torch.losses.binary_sigmoid_cross_entropy
 
 :hidden:`binary_sigmoid_cross_entropy_with_clas`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.binary_sigmoid_cross_entropy_with_clas
+.. autofunction:: texar.torch.losses.binary_sigmoid_cross_entropy_with_clas
 
 
 Policy Gradient Loss
@@ -33,11 +33,11 @@ Policy Gradient Loss
 
 :hidden:`pg_loss_with_logits`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.pg_loss_with_logits
+.. autofunction:: texar.torch.losses.pg_loss_with_logits
 
 :hidden:`pg_loss_with_log_probs`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.pg_loss_with_log_probs
+.. autofunction:: texar.torch.losses.pg_loss_with_log_probs
 
 
 Reward
@@ -45,7 +45,7 @@ Reward
 
 :hidden:`discount_reward`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.discount_reward
+.. autofunction:: texar.torch.losses.discount_reward
 
 
 Adversarial Loss
@@ -53,7 +53,7 @@ Adversarial Loss
 
 :hidden:`binary_adversarial_losses`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.binary_adversarial_losses
+.. autofunction:: texar.torch.losses.binary_adversarial_losses
 
 
 Entropy
@@ -61,11 +61,11 @@ Entropy
 
 :hidden:`entropy_with_logits`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.entropy_with_logits
+.. autofunction:: texar.torch.losses.entropy_with_logits
 
 :hidden:`sequence_entropy_with_logits`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.sequence_entropy_with_logits
+.. autofunction:: texar.torch.losses.sequence_entropy_with_logits
 
 
 Loss Utilities
@@ -73,12 +73,12 @@ Loss Utilities
 
 :hidden:`mask_and_reduce`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.mask_and_reduce
+.. autofunction:: texar.torch.losses.mask_and_reduce
 
 :hidden:`reduce_batch_time`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.reduce_batch_time
+.. autofunction:: texar.torch.losses.reduce_batch_time
 
 :hidden:`reduce_dimensions`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.losses.reduce_dimensions
+.. autofunction:: texar.torch.losses.reduce_dimensions
