+++ docs/code/utils.rst
@@ -9,16 +9,16 @@ Frequent Use
 
 :hidden:`AverageRecorder`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autoclass:: texar.utils.AverageRecorder
+.. autoclass:: texar.torch.utils.AverageRecorder
     :members:
 
 :hidden:`compat_as_text`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.compat_as_text
+.. autofunction:: texar.torch.utils.compat_as_text
 
 :hidden:`write_paired_text`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.write_paired_text
+.. autofunction:: texar.torch.utils.write_paired_text
 
 
 IO
@@ -26,11 +26,11 @@ IO
 
 :hidden:`write_paired_text`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.write_paired_text
+.. autofunction:: texar.torch.utils.write_paired_text
 
 :hidden:`maybe_create_dir`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.maybe_create_dir
+.. autofunction:: texar.torch.utils.maybe_create_dir
 
 
 DType
@@ -38,15 +38,15 @@ DType
 
 :hidden:`get_numpy_dtype`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.get_numpy_dtype
+.. autofunction:: texar.torch.utils.get_numpy_dtype
 
 :hidden:`maybe_hparams_to_dict`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.maybe_hparams_to_dict
+.. autofunction:: texar.torch.utils.maybe_hparams_to_dict
 
 :hidden:`compat_as_text`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.compat_as_text
+.. autofunction:: texar.torch.utils.compat_as_text
 
 
 Shape
@@ -54,19 +54,19 @@ Shape
 
 :hidden:`mask_sequences`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.mask_sequences
+.. autofunction:: texar.torch.utils.mask_sequences
 
 :hidden:`transpose_batch_time`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.transpose_batch_time
+.. autofunction:: texar.torch.utils.transpose_batch_time
 
 :hidden:`flatten`
 ~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.flatten
+.. autofunction:: texar.torch.utils.flatten
 
 :hidden:`pad_and_concat`
 ~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.pad_and_concat
+.. autofunction:: texar.torch.utils.pad_and_concat
 
 
 Dictionary
@@ -74,23 +74,23 @@ Dictionary
 
 :hidden:`dict_patch`
 ~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.dict_patch
+.. autofunction:: texar.torch.utils.dict_patch
 
 :hidden:`dict_lookup`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.dict_lookup
+.. autofunction:: texar.torch.utils.dict_lookup
 
 :hidden:`dict_fetch`
 ~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.dict_fetch
+.. autofunction:: texar.torch.utils.dict_fetch
 
 :hidden:`dict_pop`
 ~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.dict_pop
+.. autofunction:: texar.torch.utils.dict_pop
 
 :hidden:`flatten_dict`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.flatten_dict
+.. autofunction:: texar.torch.utils.flatten_dict
 
 
 String
@@ -98,27 +98,27 @@ String
 
 :hidden:`strip_token`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.strip_token
+.. autofunction:: texar.torch.utils.strip_token
 
 :hidden:`strip_eos`
 ~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.strip_eos
+.. autofunction:: texar.torch.utils.strip_eos
 
 :hidden:`strip_special_tokens`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.strip_special_tokens
+.. autofunction:: texar.torch.utils.strip_special_tokens
 
 :hidden:`str_join`
 ~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.str_join
+.. autofunction:: texar.torch.utils.str_join
 
 :hidden:`default_str`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.default_str
+.. autofunction:: texar.torch.utils.default_str
 
 :hidden:`uniquify_str`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.uniquify_str
+.. autofunction:: texar.torch.utils.uniquify_str
 
 
 Meta
@@ -126,47 +126,47 @@ Meta
 
 :hidden:`check_or_get_class`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.check_or_get_class
+.. autofunction:: texar.torch.utils.check_or_get_class
 
 :hidden:`get_class`
 ~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.get_class
+.. autofunction:: texar.torch.utils.get_class
 
 :hidden:`check_or_get_instance`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.check_or_get_instance
+.. autofunction:: texar.torch.utils.check_or_get_instance
 
 :hidden:`get_instance`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.get_instance
+.. autofunction:: texar.torch.utils.get_instance
 
 :hidden:`check_or_get_instance_with_redundant_kwargs`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.check_or_get_instance_with_redundant_kwargs
+.. autofunction:: texar.torch.utils.check_or_get_instance_with_redundant_kwargs
 
 :hidden:`get_instance_with_redundant_kwargs`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.get_instance_with_redundant_kwargs
+.. autofunction:: texar.torch.utils.get_instance_with_redundant_kwargs
 
 :hidden:`get_function`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.get_function
+.. autofunction:: texar.torch.utils.get_function
 
 :hidden:`call_function_with_redundant_kwargs`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.call_function_with_redundant_kwargs
+.. autofunction:: texar.torch.utils.call_function_with_redundant_kwargs
 
 :hidden:`get_args`
 ~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.get_args
+.. autofunction:: texar.torch.utils.get_args
 
 :hidden:`get_default_arg_values`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.get_default_arg_values
+.. autofunction:: texar.torch.utils.get_default_arg_values
 
 :hidden:`get_instance_kwargs`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.get_instance_kwargs
+.. autofunction:: texar.torch.utils.get_instance_kwargs
 
 
 Misc
@@ -174,23 +174,23 @@ Misc
 
 :hidden:`ceildiv`
 ~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.ceildiv
+.. autofunction:: texar.torch.utils.ceildiv
 
 :hidden:`map_structure`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.map_structure
+.. autofunction:: texar.torch.utils.map_structure
 
 :hidden:`map_structure_zip`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.map_structure_zip
+.. autofunction:: texar.torch.utils.map_structure_zip
 
 :hidden:`beam_search`
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-.. autofunction:: texar.utils.beam_search.beam_search
+.. autofunction:: texar.torch.utils.beam_search.beam_search
 
 
 
 AverageRecorder
 ==========================
-.. autoclass:: texar.utils.AverageRecorder
+.. autoclass:: texar.torch.utils.AverageRecorder
     :members:
