+++ docs/conf.py
@@ -20,8 +20,8 @@ import os
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 sys.path.insert(0, os.path.abspath('..'))
-from texar import __version__
-from texar.version import VERSION_SHORT as __version_short__
+from texar.torch import __version__
+from texar.torch.version import VERSION_SHORT as __version_short__
 
 # -- General configuration ------------------------------------------------
